services/data-pipeline/pyproject.toml
[project]
name = "finchie_data_pipeline"
version = "0.1.0"
description = "Automated ETL pipeline for extracting financial data from multiple sources, processing and loading it into Finchie's API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.165.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
]
